aircraft_api/mission_match/requirements.txt
celery==5.4.0
django-celery-beat==2.7.0
lxml==5.3.0
requests==2.32.3